requirements.txt
opencv-python
mediapipe
google-cloud-storage